requirements.txt
Django==5.1.7
Pillow==10.2.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
python-dotenv==1.0.1
django-allauth==0.61.1
django-filter==23.5
django-debug-toolbar==4.3.0
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9 
django
Pillow
django-crispy-forms
crispy-bootstrap5
python-dotenv
django-allauth
django-filter
django-debug-toolbar
gunicorn
whitenoise
psycopg2-binary
django-cloudinary-storage
django-crispy-forms
djangorestframework